requirements.txt
Django==2.0.3
pytz==2018.3
djangorestframework==3.8.0
django-rest-swagger===2.1.2